backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlmodel==0.0.14
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Audio processing
openai-whisper==20231117
# torch and torchaudio required by openai-whisper
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# AI/ML
transformers==4.30.2
sentence-transformers==2.2.2
huggingface_hub<0.16.0
langchain==0.0.335
langchain-community==0.0.6
faiss-cpu==1.7.4
numpy==1.24.3

# Google Calendar API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# Utilities
psutil==5.9.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
